feat: compare to base lcov file
